rdtech-dps: add support for RD6006P, RD6012P and RD6024
RD6012P has two current ranges, so support for getting/setting/listing
range is added. This also means that the model table is reorganized to
support multiple ranges.

In order for the range setting and the current multiplier to be up to
date, the function rdtech_dps_update_range is called when
required. This function reads the range register of the device.