korad_kaxxxxp : Add quirk for Velleman PS3005DV1.3 processing time
The Velleman PS3005D V1.3 power supply ( driver korad_kaxxxxp ) seems to
have a quirky RS232 interface, that requires waiting a longer time after
some commands before being allowed to send a new command.
Failing to respect this delay means the new command will NOT be executed.
This is especially problematic as the actual driver would report "0.000" value
for current when the power supply failed to answer the IOUT? request,
despite the output current being non zero.

This commit adds a quirk for this specific revision, and introduces a longer
delay for the affected commands ( STATUS? + any set command ).
The behavior is unchanged for other models.

Fix code style issues