rdtech-um: migrate from binary helper channel to feed queue
Use common feed queue support for the submission of analog values to the
sigrok session. This leaves the extraction of the value from binary data
images with a factor of 1 for binary helpers, but uses existing code to
scale the value (which operates at higher accuracy) and to submit values
to the sigrok session.

It's up to debate whether a scaling factor value of { 10, 1000, } can be
recognized as 10mV by readers, and how this phrase relates to the other
spec of 2 significant digits. Might be acceptable. Offsetting and MQ flags
are not used by this driver. The exponent representation was used in the
channels table because some models support 100uA resolution for current.
Engineering exponents are preferred for readability (rationals are not
"normalized" to scientific presentation with "odd" exponent values).

Register a device clear routine so that the driver releases resources
that were allocated during scan and device instance creation.