rdtech-tc: use literal string for poll command instead of array
Use a char pointer for the poll command which is a string literal. Which
eliminates the "necessity" to strip the NUL terminator from an array
after its construction. The strlen() for literals also gets resolved at
compile time. Which means that execution cost remains identical.

Rename the variable to lower case, it's not a preprocessor symbol.