rdtech-tc: migrate to feed queue helper
Implement the submission of analog data to the sigrok session by means
of the common feed queue logic. Use the feed queue's scaling support.
Implement a .dev_clear() routine which releases resources.

This exlusively leaves the extraction of numbers from binary data as
the only remaining use case for the binary helpers logic.