scpi: reduce indentation in continuation lines (function signatures)
Whitespace change exclusively, no change in behaviour. Reduce the level
of indentation of continuation lines in common SCPI code's header and
source files. See a whitespace ignoring diff for review.

A TAB is sufficient to represent a level of indentation. More than this,
up to four of them, or additional SPACE after TAB is excessive. Drop it.