scpi: add "block ends" transport method, needed for SCPI over serial
The SCPI protocol usually communicates requests (commands, queries) and
responses in ASCII format. But also supports the concept of data blocks
which carry raw byte sequences. Checking for CR or LF characters in
response text does not combine well with transmission of raw bytes.

Introduce a "block ends" method for SCPI transports. Common code is
aware when a data block is received, and where it ends. That's when
transports can re-consider their end-of-message condition for the
remaining reception of that very response. The transport method sees
all previously accumulated receive data that follows the data block.

Implement that "block ends" method in the serial transport, which
motivated the introduction of the method. Void falsely detected "is
complete" conditions that originate from seeing CR or LF in byte data.

Other transports derive their respective end-of-message condition from
other information which is not affected (EOI signals, VISA status bits,
TCP with length specs in headers before data, USB transfer flags, etc).