scpi: add "get text then block" reader (Tektronix TDS2000)
SCPI responses are expected to either come in pure text format (ASCII),
_or_ raw bytes (data blocks). Tektronix TDS2000 was observed to respond
with the unusual combination of a data block that is preceeded(!) by
variable length(!) text fields which are separated by semicolon. This
is the consequence of the "WAVF?" request being an aggregate, sending
several properties including the "CURVE" response in the same message.

Introduce a rather specific "get (an optional) text then data block"
reader routine, which uses caller provided logic to identify the
position of the data block in the response. Tested with TDS 220.

The implementation re-uses the previously existing "get block" routine
and extends it to optionally get a leading text. While lots of more
diagnostics are added. Two very thin public wrapper routines provide
the "get block" and "get (text and) block" semantics.

There is excessive diagnostics. Levels and wording needs adjustment.