tests: raise maximum version text length (dirty flag)
The previous implementation of the version text length check accounted
for the major, minor, micro numbers and the git hash. Raise the upper
limit of the accepted text length range to also account for the optional
"dirty" flag. Update comments to keep awareness of the magic numbers in
the length calculation.