greatfet: first operational GreatFET One driver implementation
This sigrok driver for the GreatFET One device as a logic analyzer is
based on https://github.com/Qyriad/libsigrok and was massaged to use
common infrastructure where appropriate, and adjust coding style to
match the sigrok.org project's code base.

This implementation supports the acquisition of 8 channels at rates up
to 40.8MHz, and higher rates up to 204MHz for reduced channel counts.
Soft triggers are not supported in this implementation, ideally common
support would cover this more transparently in the future.

Tested with v2021.2.1 firmware as available via 'pip install greatfet'.

Support for 16 channels is prepared but disabled by default. Requires
firmware support which is not officially available.